+++ modules/2d.js
@@ -1,4 +1,4 @@
-import {egal,randint,choice,rangeMinMax,unSiPositifMoinsUnSinon,arrondi,arrondi_virgule,calcul,tex_nombre,nombre_avec_espace,string_nombre,premierMultipleSuperieur,premierMultipleInferieur} from "/modules/outils.js"
+import {egal,randint,choice,rangeMinMax,unSiPositifMoinsUnSinon,arrondi,arrondi_virgule,calcul,lettre_depuis_chiffre,tex_nombre,nombre_avec_espace,string_nombre,premierMultipleSuperieur,premierMultipleInferieur} from "/modules/outils.js"
 
 /*
   MathALEA2D
@@ -496,7 +496,7 @@ export function barycentre(p, nom, positionLabel = "above") {
  *
  * @Auteur Jean-Claude Lhote
  */
-function Droite(arg1, arg2, arg3, arg4, color) {
+function Droite(arg1, arg2, arg3, arg4) {
   ObjetMathalea2D.call(this);
   if (arguments.length == 2) {
     this.nom=""
@@ -1295,712 +1295,7 @@ function Segment(arg1, arg2, arg3, arg4, color) {
     this.y2 = arg2.y;
     this.color = arg3;
   } else if (arguments.length == 4) {
-    this.x1 = arrondi(arg1,2);function Droite(arg1, arg2, arg3, arg4) {
-      ObjetMathalea2D.call(this);
-      if (arguments.length == 2) {
-        this.x1 = arg1.x;
-        this.y1 = arg1.y;
-        this.x2 = arg2.x;
-        this.y2 = arg2.y;
-        this.a = calcul(this.y1 - this.y2);
-        this.b = calcul(this.x2 - this.x1);
-        this.c = calcul(
-          (this.x1 - this.x2) * this.y1 + (this.y2 - this.y1) * this.x1
-        );
-      } else if (arguments.length == 3) {
-        if (typeof arg1 == "number") {
-          // droite d'équation ax +by +c =0
-          this.a = arg1;
-          this.b = arg2;
-          this.c = arg3;
-          a=arg1
-          b=arg2
-          c=arg3
-          if (egal(a, 0)) {
-            this.x1 = 0;
-            this.x2 = 1;
-            this.y1 = calcul(-c / b);
-            this.y2 = calcul(-c / b);
-          } else if (egal(b, 0)) {
-            this.y1 = 0;
-            this.y2 = 1;
-            this.x1 = calcul(-c / a);
-            this.x2 = calcul(-c / a);
-          } else {
-            this.x1 = 0;
-            this.y1 = calcul(-c / b);
-            this.x2 = 1;
-            this.y2 = calcul((-c - a) / b);
-          }
-        } else {
-          this.x1 = arg1.x;
-          this.y1 = arg1.y;
-          this.x2 = arg2.x;
-          this.y2 = arg2.y;
-          this.a = calcul(this.y1 - this.y2);
-          this.b = calcul(this.x2 - this.x1);
-          this.c = calcul(
-            (this.x1 - this.x2) * this.y1 + (this.y2 - this.y1) * this.x1
-          );
-          this.nom = arg3;
-        }
-      } else if (arguments.length == 4) {
-        if (typeof arg1 == "number") {
-          this.a = arg1;
-          this.b = arg2;
-          this.c = arg3;
-          a=arg1
-          b=arg2
-          c=arg3
-          nom=arg4
-          if (egal(a, 0)) {
-            this.x1 = 0;
-            this.x2 = 1;
-            this.y1 = calcul(-c / b);
-            this.y2 = calcul(-c / b);
-          } else if (egal(b, 0)) {
-            this.y1 = 0;
-            this.y2 = 1;
-            this.x1 = calcul(-c / a);
-            this.x2 = calcul(-c / a);
-          } else {
-            this.x1 = 0;
-            this.y1 = calcul(-c / b);
-            this.x2 = 1;
-            this.y2 = calcul((-c - a) / b);
-          }
-          this.nom = nom;
-        } else {
-          this.x1 = arg1.x;
-          this.y1 = arg1.y;
-          this.x2 = arg2.x;
-          this.y2 = arg2.y;
-          this.a = calcul(this.y1 - this.y2);
-          this.b = calcul(this.x2 - this.x1);
-          this.c = calcul(
-            (this.x1 - this.x2) * this.y1 + (this.y2 - this.y1) * this.x1
-          );
-          this.nom = arg3;
-          this.color = arg4;
-        }
-      }
-      if (this.b != 0) this.pente = calcul(-this.a / this.b);
-      /*	if (this.b==0) {
-        this.angleAvecHorizontale = 90
-      } else {
-        this.angleAvecHorizontale = calcul(Math.atan(this.pente)*180/Math.PI,1)
-    
-      }
-      */
-      this.normal = vecteur(this.a, this.b);
-      this.directeur = vecteur(this.b,- this.a);
-      this.angleAvecHorizontale = angleOriente(
-        point(1, 0),
-        point(0, 0),
-        point(this.directeur.x, this.directeur.y)
-      );
-      let absNom,ordNom,leNom
-      let pointXmin=pointSurDroite(this,fenetreMathalea2d[0])
-      if (this.nom!='') {
-        pointXmin=pointSurDroite(this,fenetreMathalea2d[0])
-        if (pointXmin.y>fenetreMathalea2d[1]&&pointXmin.y<fenetreMathalea2d[3]) {
-          absNom=fenetreMathalea2d[0]+1
-          ordNom=pointXmin.y+0.5
-        }
-        else {
-          pointXmin=pointSurDroite(this,fenetreMathalea2d[2])
-          if (pointXmin.y>fenetreMathalea2d[1]&&pointXmin.y<fenetreMathalea2d[3]) {
-            absNom=fenetreMathalea2d[2]-1
-            ordNom=pointXmin.y+0.5
-          }
-          else {
-            pointXmin=pointIntersectionDD(this,droiteHorizontaleParPoint(point(0,fenetreMathalea2d[1])))
-            if (pointXmin.x>fenetreMathalea2d[0]&&pointXmin.x<fenetreMathalea2d[2]) {
-              absNom=pointXmin.x+0.5
-              ordNom=fenetreMathalea2d[1]+1
-            }
-            else {
-              pointXmin=pointIntersectionDD(this,droiteHorizontaleParPoint(point(0,fenetreMathalea2d[3])))
-              if (pointXmin.x>fenetreMathalea2d[0]&&pointXmin.x<fenetreMathalea2d[2]) {
-                absNom=pointXmin.x+0.5
-                ordNom=fenetreMathalea2d[3]-1
-              }
-              else {
-                absNom=(fenetreMathalea2d[0]+fenetreMathalea2d[2]/2)
-                ordNom=pointSurDroite(this,absNom).y+0.5
-              }
-            }
-          }
-        }
-        leNom=texteParPosition(this.nom,absNom,ordNom,"milieu",this.color,1,"milieu",true)
-    
-      }
-      this.svg = function (coeff) {
-       
-        if (this.epaisseur != 1) {
-          this.style += ` stroke-width="${this.epaisseur}" `;
-        }
-        if (Boolean(this.pointilles)) {
-          switch (this.pointilles) {
-            case 1 :
-              this.style += ` stroke-dasharray="6 10" `;
-              break;
-            case 2 : 
-            this.style += ` stroke-dasharray="6 3" `;
-            break;       
-            case 3 :
-              this.style += ` stroke-dasharray="3 2 6 2 " `;
-              break;      
-            default : 
-            this.style += ` stroke-dasharray="5 5" `;
-            break; 
-          }
-    
-        }function Droite(arg1, arg2, arg3, arg4) {
-          ObjetMathalea2D.call(this);
-          if (arguments.length == 2) {
-            this.x1 = arg1.x;
-            this.y1 = arg1.y;
-            this.x2 = arg2.x;
-            this.y2 = arg2.y;
-            this.a = calcul(this.y1 - this.y2);
-            this.b = calcul(this.x2 - this.x1);
-            this.c = calcul(
-              (this.x1 - this.x2) * this.y1 + (this.y2 - this.y1) * this.x1
-            );
-          } else if (arguments.length == 3) {
-            if (typeof arg1 == "number") {
-              // droite d'équation ax +by +c =0
-              this.a = arg1;
-              this.b = arg2;
-              this.c = arg3;
-              a=arg1
-              b=arg2
-              c=arg3
-              if (egal(a, 0)) {
-                this.x1 = 0;
-                this.x2 = 1;
-                this.y1 = calcul(-c / b);
-                this.y2 = calcul(-c / b);
-              } else if (egal(b, 0)) {
-                this.y1 = 0;
-                this.y2 = 1;
-                this.x1 = calcul(-c / a);
-                this.x2 = calcul(-c / a);
-              } else {
-                this.x1 = 0;
-                this.y1 = calcul(-c / b);
-                this.x2 = 1;
-                this.y2 = calcul((-c - a) / b);
-              }
-            } else {
-              this.x1 = arg1.x;
-              this.y1 = arg1.y;
-              this.x2 = arg2.x;
-              this.y2 = arg2.y;
-              this.a = calcul(this.y1 - this.y2);
-              this.b = calcul(this.x2 - this.x1);
-              this.c = calcul(
-                (this.x1 - this.x2) * this.y1 + (this.y2 - this.y1) * this.x1
-              );
-              this.nom = arg3;
-            }
-          } else if (arguments.length == 4) {
-            if (typeof arg1 == "number") {
-              this.a = arg1;
-              this.b = arg2;
-              this.c = arg3;
-              a=arg1
-              b=arg2
-              c=arg3
-              nom=arg4
-              if (egal(a, 0)) {
-                this.x1 = 0;
-                this.x2 = 1;
-                this.y1 = calcul(-c / b);
-                this.y2 = calcul(-c / b);
-              } else if (egal(b, 0)) {
-                this.y1 = 0;
-                this.y2 = 1;
-                this.x1 = calcul(-c / a);
-                this.x2 = calcul(-c / a);
-              } else {
-                this.x1 = 0;
-                this.y1 = calcul(-c / b);
-                this.x2 = 1;
-                this.y2 = calcul((-c - a) / b);
-              }
-              this.nom = nom;
-            } else {
-              this.x1 = arg1.x;
-              this.y1 = arg1.y;
-              this.x2 = arg2.x;
-              this.y2 = arg2.y;
-              this.a = calcul(this.y1 - this.y2);
-              this.b = calcul(this.x2 - this.x1);
-              this.c = calcul(
-                (this.x1 - this.x2) * this.y1 + (this.y2 - this.y1) * this.x1
-              );
-              this.nom = arg3;
-              this.color = arg4;
-            }
-          }
-          if (this.b != 0) this.pente = calcul(-this.a / this.b);
-          /*	if (this.b==0) {
-            this.angleAvecHorizontale = 90
-          } else {
-            this.angleAvecHorizontale = calcul(Math.atan(this.pente)*180/Math.PI,1)
-        
-          }
-          */
-          this.normal = vecteur(this.a, this.b);
-          this.directeur = vecteur(this.b,- this.a);
-          this.angleAvecHorizontale = angleOriente(
-            point(1, 0),
-            point(0, 0),
-            point(this.directeur.x, this.directeur.y)
-          );
-          let absNom,ordNom,leNom
-          let pointXmin=pointSurDroite(this,fenetreMathalea2d[0])
-          if (this.nom!='') {
-            pointXmin=pointSurDroite(this,fenetreMathalea2d[0])
-            if (pointXmin.y>fenetreMathalea2d[1]&&pointXmin.y<fenetreMathalea2d[3]) {
-              absNom=fenetreMathalea2d[0]+1
-              ordNom=pointXmin.y+0.5
-            }
-            else {
-              pointXmin=pointSurDroite(this,fenetreMathalea2d[2])
-              if (pointXmin.y>fenetreMathalea2d[1]&&pointXmin.y<fenetreMathalea2d[3]) {
-                absNom=fenetreMathalea2d[2]-1
-                ordNom=pointXmin.y+0.5
-              }
-              else {
-                pointXmin=pointIntersectionDD(this,droiteHorizontaleParPoint(point(0,fenetreMathalea2d[1])))
-                if (pointXmin.x>fenetreMathalea2d[0]&&pointXmin.x<fenetreMathalea2d[2]) {
-                  absNom=pointXmin.x+0.5
-                  ordNom=fenetreMathalea2d[1]+1
-                }
-                else {
-                  pointXmin=pointIntersectionDD(this,droiteHorizontaleParPoint(point(0,fenetreMathalea2d[3])))
-                  if (pointXmin.x>fenetreMathalea2d[0]&&pointXmin.x<fenetreMathalea2d[2]) {
-                    absNom=pointXmin.x+0.5
-                    ordNom=fenetreMathalea2d[3]-1
-                  }
-                  else {
-                    absNom=(fenetreMathalea2d[0]+fenetreMathalea2d[2]/2)
-                    ordNom=pointSurDroite(this,absNom).y+0.5
-                  }
-                }
-              }
-            }
-            leNom=texteParPosition(this.nom,absNom,ordNom,"milieu",this.color,1,"milieu",true)
-        
-          }
-          this.svg = function (coeff) {
-           
-            if (this.epaisseur != 1) {
-              this.style += ` stroke-width="${this.epaisseur}" `;
-            }
-            if (Boolean(this.pointilles)) {
-              switch (this.pointilles) {
-                case 1 :
-                  this.style += ` stroke-dasharray="6 10" `;
-                  break;
-                case 2 : 
-                this.style += ` stroke-dasharray="6 3" `;
-                break;       
-                case 3 :
-                  this.style += ` stroke-dasharray="3 2 6 2 " `;
-                  break;      
-                default : 
-                this.style += ` stroke-dasharray="5 5" `;
-                break; 
-              }
-        
-            }
-            if (this.opacite != 1) {
-              this.style += ` stroke-opacity="${this.opacite}" `;
-            }
-            let A = point(this.x1, this.y1);
-            let B = point(this.x2, this.y2);
-            let A1 = pointSurSegment(A, B, -50);
-            let B1 = pointSurSegment(B, A, -50);
-            if (typeof(leNom)=='undefined')
-               return `<line x1="${A1.xSVG(coeff)}" y1="${A1.ySVG(coeff)}" x2="${B1.xSVG(
-              coeff
-            )}" y2="${B1.ySVG(coeff)}" stroke="${this.color}" ${this.style} id ="${this.id}" />`;
-            else return `<line x1="${A1.xSVG(coeff)}" y1="${A1.ySVG(coeff)}" x2="${B1.xSVG(
-              coeff
-            )}" y2="${B1.ySVG(coeff)}" stroke="${this.color}" ${this.style} id ="${this.id}" />`+leNom.svg(coeff);
-          };
-        
-          this.tikz = function () {
-            let tableauOptions = [];
-            if (this.color.length > 1 && this.color !== "black") {
-              tableauOptions.push(this.color);
-            }
-            if (this.epaisseur != 1) {
-              tableauOptions.push(`line width = ${this.epaisseur}`);
-            }
-            if (Boolean(this.pointilles)) {
-             switch (this.pointilles) {
-                case 1 :
-                  tableauOptions.push(` dash dot `);
-                  break;
-                case 2 : 
-                tableauOptions.push(` dash dash dot `);
-                break;       
-                case 3 :
-                  tableauOptions.push(` dash dot dot `);
-                  break;      
-                default : 
-                  tableauOptions.push(` dashed `);
-                break; 
-              }
-            }
-        
-            if (this.opacite != 1) {
-              tableauOptions.push(`opacity = ${this.opacite}`);
-            }
-        
-            let optionsDraw = [];
-            if (tableauOptions.length > 0) {
-              optionsDraw = "[" + tableauOptions.join(",") + "]";
-            }
-            let A = point(this.x1, this.y1);
-            let B = point(this.x2, this.y2);
-            let A1 = pointSurSegment(A, B, -10);
-            let B1 = pointSurSegment(B, A, -10);
-            return `\\draw${optionsDraw} (${A1.x},${A1.y})--(${B1.x},${B1.y});`+leNom.tikz();
-          };
-          this.svgml = function(coeff,amp){
-            let A = point(this.x1, this.y1);
-            let B = point(this.x2, this.y2);
-            let A1 = pointSurSegment(A, B, -10);
-            let B1 = pointSurSegment(B, A, -10);
-            let s=segment(A1,B1,this.color)
-            s.isVisible=false
-          return s.svgml(coeff,amp)+leNom.svg(coeff)
-          }
-          this.tikzml = function(amp){
-            let A = point(this.x1, this.y1);
-            let B = point(this.x2, this.y2);
-            let A1 = pointSurSegment(A, B, -10);
-            let B1 = pointSurSegment(B, A, -10);
-            let s=segment(A1,B1,this.color)
-            s.isVisible=false
-          return s.tikzml(amp)+leNom.tikz()
-          }
-        }function Droite(arg1, arg2, arg3, arg4) {
-          ObjetMathalea2D.call(this);
-          if (arguments.length == 2) {
-            this.x1 = arg1.x;
-            this.y1 = arg1.y;
-            this.x2 = arg2.x;
-            this.y2 = arg2.y;
-            this.a = calcul(this.y1 - this.y2);
-            this.b = calcul(this.x2 - this.x1);
-            this.c = calcul(
-              (this.x1 - this.x2) * this.y1 + (this.y2 - this.y1) * this.x1
-            );
-          } else if (arguments.length == 3) {
-            if (typeof arg1 == "number") {
-              // droite d'équation ax +by +c =0
-              this.a = arg1;
-              this.b = arg2;
-              this.c = arg3;
-              a=arg1
-              b=arg2
-              c=arg3
-              if (egal(a, 0)) {
-                this.x1 = 0;
-                this.x2 = 1;
-                this.y1 = calcul(-c / b);
-                this.y2 = calcul(-c / b);
-              } else if (egal(b, 0)) {
-                this.y1 = 0;
-                this.y2 = 1;
-                this.x1 = calcul(-c / a);
-                this.x2 = calcul(-c / a);
-              } else {
-                this.x1 = 0;
-                this.y1 = calcul(-c / b);
-                this.x2 = 1;
-                this.y2 = calcul((-c - a) / b);
-              }
-            } else {
-              this.x1 = arg1.x;
-              this.y1 = arg1.y;
-              this.x2 = arg2.x;
-              this.y2 = arg2.y;
-              this.a = calcul(this.y1 - this.y2);
-              this.b = calcul(this.x2 - this.x1);
-              this.c = calcul(
-                (this.x1 - this.x2) * this.y1 + (this.y2 - this.y1) * this.x1
-              );
-              this.nom = arg3;
-            }
-          } else if (arguments.length == 4) {
-            if (typeof arg1 == "number") {
-              this.a = arg1;
-              this.b = arg2;
-              this.c = arg3;
-              a=arg1
-              b=arg2
-              c=arg3
-              nom=arg4
-              if (egal(a, 0)) {
-                this.x1 = 0;
-                this.x2 = 1;
-                this.y1 = calcul(-c / b);
-                this.y2 = calcul(-c / b);
-              } else if (egal(b, 0)) {
-                this.y1 = 0;
-                this.y2 = 1;
-                this.x1 = calcul(-c / a);
-                this.x2 = calcul(-c / a);
-              } else {
-                this.x1 = 0;
-                this.y1 = calcul(-c / b);
-                this.x2 = 1;
-                this.y2 = calcul((-c - a) / b);
-              }
-              this.nom = nom;
-            } else {
-              this.x1 = arg1.x;
-              this.y1 = arg1.y;
-              this.x2 = arg2.x;
-              this.y2 = arg2.y;
-              this.a = calcul(this.y1 - this.y2);
-              this.b = calcul(this.x2 - this.x1);
-              this.c = calcul(
-                (this.x1 - this.x2) * this.y1 + (this.y2 - this.y1) * this.x1
-              );
-              this.nom = arg3;
-              this.color = arg4;
-            }
-          }
-          if (this.b != 0) this.pente = calcul(-this.a / this.b);
-          /*	if (this.b==0) {
-            this.angleAvecHorizontale = 90
-          } else {
-            this.angleAvecHorizontale = calcul(Math.atan(this.pente)*180/Math.PI,1)
-        
-          }
-          */
-          this.normal = vecteur(this.a, this.b);
-          this.directeur = vecteur(this.b,- this.a);
-          this.angleAvecHorizontale = angleOriente(
-            point(1, 0),
-            point(0, 0),
-            point(this.directeur.x, this.directeur.y)
-          );
-          let absNom,ordNom,leNom
-          let pointXmin=pointSurDroite(this,fenetreMathalea2d[0])
-          if (this.nom!='') {
-            pointXmin=pointSurDroite(this,fenetreMathalea2d[0])
-            if (pointXmin.y>fenetreMathalea2d[1]&&pointXmin.y<fenetreMathalea2d[3]) {
-              absNom=fenetreMathalea2d[0]+1
-              ordNom=pointXmin.y+0.5
-            }
-            else {
-              pointXmin=pointSurDroite(this,fenetreMathalea2d[2])
-              if (pointXmin.y>fenetreMathalea2d[1]&&pointXmin.y<fenetreMathalea2d[3]) {
-                absNom=fenetreMathalea2d[2]-1
-                ordNom=pointXmin.y+0.5
-              }
-              else {
-                pointXmin=pointIntersectionDD(this,droiteHorizontaleParPoint(point(0,fenetreMathalea2d[1])))
-                if (pointXmin.x>fenetreMathalea2d[0]&&pointXmin.x<fenetreMathalea2d[2]) {
-                  absNom=pointXmin.x+0.5
-                  ordNom=fenetreMathalea2d[1]+1
-                }
-                else {
-                  pointXmin=pointIntersectionDD(this,droiteHorizontaleParPoint(point(0,fenetreMathalea2d[3])))
-                  if (pointXmin.x>fenetreMathalea2d[0]&&pointXmin.x<fenetreMathalea2d[2]) {
-                    absNom=pointXmin.x+0.5
-                    ordNom=fenetreMathalea2d[3]-1
-                  }
-                  else {
-                    absNom=(fenetreMathalea2d[0]+fenetreMathalea2d[2]/2)
-                    ordNom=pointSurDroite(this,absNom).y+0.5
-                  }
-                }
-              }
-            }
-            leNom=texteParPosition(this.nom,absNom,ordNom,"milieu",this.color,1,"milieu",true)
-        
-          }
-          this.svg = function (coeff) {
-           
-            if (this.epaisseur != 1) {
-              this.style += ` stroke-width="${this.epaisseur}" `;
-            }
-            if (Boolean(this.pointilles)) {
-              switch (this.pointilles) {
-                case 1 :
-                  this.style += ` stroke-dasharray="6 10" `;
-                  break;
-                case 2 : 
-                this.style += ` stroke-dasharray="6 3" `;
-                break;       
-                case 3 :
-                  this.style += ` stroke-dasharray="3 2 6 2 " `;
-                  break;      
-                default : 
-                this.style += ` stroke-dasharray="5 5" `;
-                break; 
-              }
-        
-            }
-            if (this.opacite != 1) {
-              this.style += ` stroke-opacity="${this.opacite}" `;
-            }
-            let A = point(this.x1, this.y1);
-            let B = point(this.x2, this.y2);
-            let A1 = pointSurSegment(A, B, -50);
-            let B1 = pointSurSegment(B, A, -50);
-            if (typeof(leNom)=='undefined')
-               return `<line x1="${A1.xSVG(coeff)}" y1="${A1.ySVG(coeff)}" x2="${B1.xSVG(
-              coeff
-            )}" y2="${B1.ySVG(coeff)}" stroke="${this.color}" ${this.style} id ="${this.id}" />`;
-            else return `<line x1="${A1.xSVG(coeff)}" y1="${A1.ySVG(coeff)}" x2="${B1.xSVG(
-              coeff
-            )}" y2="${B1.ySVG(coeff)}" stroke="${this.color}" ${this.style} id ="${this.id}" />`+leNom.svg(coeff);
-          };
-        
-          this.tikz = function () {
-            let tableauOptions = [];
-            if (this.color.length > 1 && this.color !== "black") {
-              tableauOptions.push(this.color);
-            }
-            if (this.epaisseur != 1) {
-              tableauOptions.push(`line width = ${this.epaisseur}`);
-            }
-            if (Boolean(this.pointilles)) {
-             switch (this.pointilles) {
-                case 1 :
-                  tableauOptions.push(` dash dot `);
-                  break;
-                case 2 : 
-                tableauOptions.push(` dash dash dot `);
-                break;       
-                case 3 :
-                  tableauOptions.push(` dash dot dot `);
-                  break;      
-                default : 
-                  tableauOptions.push(` dashed `);
-                break; 
-              }
-            }
-        
-            if (this.opacite != 1) {
-              tableauOptions.push(`opacity = ${this.opacite}`);
-            }
-        
-            let optionsDraw = [];
-            if (tableauOptions.length > 0) {
-              optionsDraw = "[" + tableauOptions.join(",") + "]";
-            }
-            let A = point(this.x1, this.y1);
-            let B = point(this.x2, this.y2);
-            let A1 = pointSurSegment(A, B, -10);
-            let B1 = pointSurSegment(B, A, -10);
-            return `\\draw${optionsDraw} (${A1.x},${A1.y})--(${B1.x},${B1.y});`+leNom.tikz();
-          };
-          this.svgml = function(coeff,amp){
-            let A = point(this.x1, this.y1);
-            let B = point(this.x2, this.y2);
-            let A1 = pointSurSegment(A, B, -10);
-            let B1 = pointSurSegment(B, A, -10);
-            let s=segment(A1,B1,this.color)
-            s.isVisible=false
-          return s.svgml(coeff,amp)+leNom.svg(coeff)
-          }
-          this.tikzml = function(amp){
-            let A = point(this.x1, this.y1);
-            let B = point(this.x2, this.y2);
-            let A1 = pointSurSegment(A, B, -10);
-            let B1 = pointSurSegment(B, A, -10);
-            let s=segment(A1,B1,this.color)
-            s.isVisible=false
-          return s.tikzml(amp)+leNom.tikz()
-          }
-        }
-        if (this.opacite != 1) {
-          this.style += ` stroke-opacity="${this.opacite}" `;
-        }
-        let A = point(this.x1, this.y1);
-        let B = point(this.x2, this.y2);
-        let A1 = pointSurSegment(A, B, -50);
-        let B1 = pointSurSegment(B, A, -50);
-        if (typeof(leNom)=='undefined')
-           return `<line x1="${A1.xSVG(coeff)}" y1="${A1.ySVG(coeff)}" x2="${B1.xSVG(
-          coeff
-        )}" y2="${B1.ySVG(coeff)}" stroke="${this.color}" ${this.style} id ="${this.id}" />`;
-        else return `<line x1="${A1.xSVG(coeff)}" y1="${A1.ySVG(coeff)}" x2="${B1.xSVG(
-          coeff
-        )}" y2="${B1.ySVG(coeff)}" stroke="${this.color}" ${this.style} id ="${this.id}" />`+leNom.svg(coeff);
-      };
-    
-      this.tikz = function () {
-        let tableauOptions = [];
-        if (this.color.length > 1 && this.color !== "black") {
-          tableauOptions.push(this.color);
-        }
-        if (this.epaisseur != 1) {
-          tableauOptions.push(`line width = ${this.epaisseur}`);
-        }
-        if (Boolean(this.pointilles)) {
-         switch (this.pointilles) {
-            case 1 :
-              tableauOptions.push(` dash dot `);
-              break;
-            case 2 : 
-            tableauOptions.push(` dash dash dot `);
-            break;       
-            case 3 :
-              tableauOptions.push(` dash dot dot `);
-              break;      
-            default : 
-              tableauOptions.push(` dashed `);
-            break; 
-          }
-        }
-    
-        if (this.opacite != 1) {
-          tableauOptions.push(`opacity = ${this.opacite}`);
-        }
-    
-        let optionsDraw = [];
-        if (tableauOptions.length > 0) {
-          optionsDraw = "[" + tableauOptions.join(",") + "]";
-        }
-        let A = point(this.x1, this.y1);
-        let B = point(this.x2, this.y2);
-        let A1 = pointSurSegment(A, B, -10);
-        let B1 = pointSurSegment(B, A, -10);
-        return `\\draw${optionsDraw} (${A1.x},${A1.y})--(${B1.x},${B1.y});`+leNom.tikz();
-      };
-      this.svgml = function(coeff,amp){
-        let A = point(this.x1, this.y1);
-        let B = point(this.x2, this.y2);
-        let A1 = pointSurSegment(A, B, -10);
-        let B1 = pointSurSegment(B, A, -10);
-        let s=segment(A1,B1,this.color)
-        s.isVisible=false
-      return s.svgml(coeff,amp)+leNom.svg(coeff)
-      }
-      this.tikzml = function(amp){
-        let A = point(this.x1, this.y1);
-        let B = point(this.x2, this.y2);
-        let A1 = pointSurSegment(A, B, -10);
-        let B1 = pointSurSegment(B, A, -10);
-        let s=segment(A1,B1,this.color)
-        s.isVisible=false
-      return s.tikzml(amp)+leNom.tikz()
-      }
-    }
+    this.x1 = arrondi(arg1,2);
     this.y1 = arrondi(arg2,2);
     this.x2 = arrondi(arg3,2);
     this.y2 = arrondi(arg4,2);
@@ -5610,9 +4905,9 @@ function Grille(
   ObjetMathalea2D.call(this);
   this.color = color;
   this.opacite = opacite;
-  let objets = [];
+  let objets = [],s;
   for (let i = xmin; i <= xmax; i += step) {
-    let s = segment(i, ymin, i, ymax);
+    s = segment(i, ymin, i, ymax);
     s.color = this.color;
     s.opacite = this.opacite;
     if (pointilles) {
@@ -5621,7 +4916,7 @@ function Grille(
     objets.push(s);
   }
   for (let i = ymin; i <= ymax; i += step) {
-    let s = segment(xmin, i, xmax, i);
+    s = segment(xmin, i, xmax, i);
     s.color = this.color;
     s.opacite = this.opacite;
     if (pointilles) {
@@ -7716,9 +7011,6 @@ function TranslationInstrument(instrument,A,B,begin=0,id){
   }
 }
 
-function translationInstrument(...args){
-  return new TranslationInstrument(...args)
-}
 
 // export function deplaceInstrument(instrument, B, begin=0, id){
 //   let A = point(instrument.x,instrument.y);
@@ -7924,7 +7216,7 @@ function Labyrinthe() {
 	this.chemin = []
 	this.niveau = 3
 	this.nombres = [[]]
-  let  s1, s2, s3, s4, s5, couleur = 'brown'
+  let  couleur = 'brown'
   let chemins = [
 		[[1, 0], [2, 0], [3, 0], [4, 0], [5, 0], [5, 1], [6, 1]],
 		[[1, 0], [2, 0], [3, 0], [4, 0], [4, 1], [5, 1], [6, 1]],
@@ -8119,8 +7411,8 @@ function Labyrinthe() {
 	// Retourne le tableau d'objets du chemin
 	this.traceChemin = function (monchemin) {
 		let y = monchemin[0][1]
-		let x = 0, chemin2d = [],s1,s2
-		for (let j = 0; j < monchemin.length; j++) {
+    let x = 0, chemin2d = [],s1	
+    for (let j = 0; j < monchemin.length; j++) {
 			s1 = segment(point(x * 3 - 1.5, y * 3 + 2.5), point(monchemin[j][0] * 3 - 1.5, monchemin[j][1] * 3 + 2.5), couleur)
 			s1.pointilles = true
 			s1.stylePointilles = 2
